+++ internal/controller/core/site_controller_connect.go
@@ -71,13 +71,12 @@ func (r *SiteReconciler) reconcileConnect(
 				},
 				Server: &v1beta1.ConnectServerConfig{
 					// This will be filled in by the controller... see "Url" below
-					Address:                "",
-					FrameOptionsDashboard:  "NONE",
-					FrameOptionsContent:    "NONE",
-					DefaultContentListView: v1beta1.ContentListViewCompact,
-					LoggedInWarning:        site.Spec.Connect.LoggedInWarning,
-					PublicWarning:          site.Spec.Connect.PublicWarning,
-					HideEmailAddresses:     true,
+					Address:               "",
+					FrameOptionsDashboard: "NONE",
+					FrameOptionsContent:   "NONE",
+					LoggedInWarning:       site.Spec.Connect.LoggedInWarning,
+					PublicWarning:         site.Spec.Connect.PublicWarning,
+					HideEmailAddresses:    true,
 				},
 				Http: &v1beta1.ConnectHttpConfig{
 					ForceSecure: true,
@@ -147,6 +146,7 @@ func (r *SiteReconciler) reconcileConnect(
 			OffHostExecution:        true,
 			AdditionalRuntimeImages: site.Spec.Connect.AdditionalRuntimeImages,
 			Auth:                    site.Spec.Connect.Auth,
+			RegisterOnFirstLogin:    site.Spec.Connect.RegisterOnFirstLogin,
 			Secret:                  site.Spec.Secret,
 			WorkloadSecret:          site.Spec.WorkloadSecret,
 			Debug:                   connectDebugLog,
@@ -217,6 +217,9 @@ func (r *SiteReconciler) reconcileConnect(
 			targetConnect.Spec.ChronicleSidecarProductApiKeyEnabled = true
 		}
 	}
+
+	// Propagate additional config if configured
+	targetConnect.Spec.Config.AdditionalConfig = site.Spec.Connect.AdditionalConfig
 
 	// if volumeSource.type is set, then force volume creation for Connect
 	if site.Spec.VolumeSource.Type != v1beta1.VolumeSourceTypeNone {