+++ internal/controller/core/site_controller_workbench.go
@@ -397,11 +397,21 @@ func (r *SiteReconciler) reconcileWorkbench(
 		targetWorkbench.Spec.Config.WorkbenchIniConfig.Jupyter = site.Spec.Workbench.JupyterConfig
 	}
 
+	// Propagate additional configs for server config files
+	if site.Spec.Workbench.AdditionalConfigs != nil {
+		targetWorkbench.Spec.Config.WorkbenchIniConfig.AdditionalConfigs = site.Spec.Workbench.AdditionalConfigs
+	}
+
+	// Propagate additional configs for session config files
+	if site.Spec.Workbench.AdditionalSessionConfigs != nil {
+		targetWorkbench.Spec.Config.WorkbenchSessionIniConfig.AdditionalConfigs = site.Spec.Workbench.AdditionalSessionConfigs
+	}
+
 	// Propagate audited jobs configuration
 	if site.Spec.Workbench.AuditedJobs != nil {
 		aj := site.Spec.Workbench.AuditedJobs
 		if aj.Enabled != nil {
-			targetWorkbench.Spec.Config.RServer.AuditedJobs = *aj.Enabled
+			targetWorkbench.Spec.Config.RServer.AuditedJobs = aj.Enabled
 		}
 		if aj.StoragePath != "" {
 			targetWorkbench.Spec.Config.RServer.AuditedJobsStoragePath = aj.StoragePath
@@ -413,19 +423,19 @@ func (r *SiteReconciler) reconcileWorkbench(
 			targetWorkbench.Spec.Config.RServer.AuditedJobsPublicKeyPaths = aj.PublicKeyPaths
 		}
 		if aj.LogLimit != nil {
-			targetWorkbench.Spec.Config.RServer.AuditedJobsLogLimit = *aj.LogLimit
+			targetWorkbench.Spec.Config.RServer.AuditedJobsLogLimit = aj.LogLimit
 		}
 		if aj.DeletionExpiry != nil {
-			targetWorkbench.Spec.Config.RServer.AuditedJobsDeletionExpiry = *aj.DeletionExpiry
+			targetWorkbench.Spec.Config.RServer.AuditedJobsDeletionExpiry = aj.DeletionExpiry
 		}
 		if aj.VanillaRequired != nil {
-			targetWorkbench.Spec.Config.RServer.AuditedJobsVanillaRequired = *aj.VanillaRequired
+			targetWorkbench.Spec.Config.RServer.AuditedJobsVanillaRequired = aj.VanillaRequired
 		}
 		if aj.DetailsEnvironment != nil {
-			targetWorkbench.Spec.Config.RServer.AuditedJobsDetailsEnvironment = *aj.DetailsEnvironment
+			targetWorkbench.Spec.Config.RServer.AuditedJobsDetailsEnvironment = aj.DetailsEnvironment
 		}
 		if aj.DetailsUserDefined != nil {
-			targetWorkbench.Spec.Config.RServer.AuditedJobsDetailsUserDefined = *aj.DetailsUserDefined
+			targetWorkbench.Spec.Config.RServer.AuditedJobsDetailsUserDefined = aj.DetailsUserDefined
 		}
 	}
 
@@ -467,8 +477,8 @@ func (r *SiteReconciler) reconcileWorkbench(
 	return nil
 }
 
-func defaultWorkbenchResourceProfiles() map[string]*v1beta1.WorkbenchLauncherKubnernetesResourcesConfigSection {
-	return map[string]*v1beta1.WorkbenchLauncherKubnernetesResourcesConfigSection{
+func defaultWorkbenchResourceProfiles() map[string]*v1beta1.WorkbenchLauncherKubernetesResourcesConfigSection {
+	return map[string]*v1beta1.WorkbenchLauncherKubernetesResourcesConfigSection{
 		"default": {
 			Name:  "Small",
 			Cpus:  "1",
@@ -488,7 +498,7 @@ func defaultWorkbenchResourceProfiles() map[string]*v1beta1.WorkbenchLauncherKub
 }
 
 // getResourceProfileKeys extracts the keys from a resource profiles map
-func getResourceProfileKeys(resourceProfiles map[string]*v1beta1.WorkbenchLauncherKubnernetesResourcesConfigSection) []string {
+func getResourceProfileKeys(resourceProfiles map[string]*v1beta1.WorkbenchLauncherKubernetesResourcesConfigSection) []string {
 	keys := make([]string, 0, len(resourceProfiles))
 	for key := range resourceProfiles {
 		keys = append(keys, key)